[기능] PlayerSaveData Class 추가
[수정] PlayerStat - _currentShield 변수 float -> int 타입 수정

+++ Assets/Scripts/Player/PlayerStat.cs
@@ -10,7 +10,7 @@ public class PlayerStat : Stat
     [SerializeField] float _currentExp;
     [SerializeField] float _maxSoul;
     [SerializeField] float _currentSoul;
-    [SerializeField] float _currentShield;
+    [SerializeField] int _currentShield;
 
     
 
@@ -27,7 +27,7 @@ public class PlayerStat : Stat
     public float CurrentExp { get { return _currentExp; } set {  _currentExp = value; } }
     public float MaxSoul { get { return _maxSoul; }}
     public float CurrentSoul { get { return _currentSoul; }set { _currentSoul = value; } }
-    public float CurrentShield { get { return _currentShield; } set { _currentShield = value; } }
+    public int CurrentShield { get { return _currentShield; } set { _currentShield = value; } }
 
     public float TotalMaxHp { get { return _maxHp + _additionalHp; } }
     public float TotalAttack { get { return _currentAttack + _weaponDamage; } }
